pyproject.toml
[project]
name = "dataprocessing-example-deid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "xlsxwriter>=3.2.9",
]
